#Review Comments - Error Handler in Domain Layer
