--- app/src/main/java/com/srdroid/memedb/domain/error/ErrorHandler.kt
@@ -1,8 +1,0 @@
-package com.srdroid.memedb.domain.error
-
-import com.srdroid.memedb.data.error.ErrorEntity
-
-interface ErrorHandler {
-
-    fun getError(throwable: Throwable): ErrorEntity
-}
