+++ app/src/main/java/com/srdroid/memedb/domain/errorhandler/ErrorHandler.kt
@@ -1,0 +1,6 @@
+package com.srdroid.memedb.domain.errorhandler
+
+interface ErrorHandler {
+
+    fun getError(throwable: Throwable): ErrorEntity
+}
