+++ app/src/main/java/com/srdroid/memedb/domain/errorhandler/GeneralErrorHandlerImpl.kt
@@ -1,6 +1,5 @@
-package com.srdroid.memedb.data.error
+package com.srdroid.memedb.domain.errorhandler
 
-import com.srdroid.memedb.domain.error.ErrorHandler
 import okio.IOException
 import retrofit2.HttpException
 import java.net.HttpURLConnection
