+++ app/src/main/java/com/srdroid/memedb/domain/use_case/GetMemeDetailsUseCase.kt
@@ -2,7 +2,7 @@ package com.srdroid.memedb.domain.use_case
 
 import com.srdroid.memedb.core.AppConstants
 import com.srdroid.memedb.core.Resource
-import com.srdroid.memedb.domain.error.ErrorHandler
+import com.srdroid.memedb.domain.errorhandler.ErrorHandler
 import com.srdroid.memedb.domain.mappers.MemeModelMapper
 import com.srdroid.memedb.domain.model.MemeModel
 import com.srdroid.memedb.domain.repository.MemeDetailsRepository
