+++ app/src/test/java/com/srdroid/memedb/usecase/MemeDetailsUseCaseUT.kt
@@ -3,7 +3,7 @@ package com.srdroid.memedb.usecase
 import com.srdroid.memedb.core.ID
 import com.srdroid.memedb.core.MockResponse.getMemesModel
 import com.srdroid.memedb.core.TestCoroutineRule
-import com.srdroid.memedb.data.error.GeneralErrorHandlerImpl
+import com.srdroid.memedb.domain.errorhandler.GeneralErrorHandlerImpl
 import com.srdroid.memedb.data.model.MemeDTO
 import com.srdroid.memedb.data.repository.MemeDetailsRepositoryImpl
 import com.srdroid.memedb.domain.mappers.MemeModelMapper
